+++ dhall/tests/Dhall/Test/Tags.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE OverloadedStrings #-}
 
-module Dhall.Test.ETags where
+module Dhall.Test.Tags where
 
 import Data.Monoid ((<>))
 import Data.Text (Text)
@@ -11,30 +11,30 @@ import Turtle (FilePath)
 
 import qualified Data.Text        as Text
 import qualified Data.Text.IO     as Text.IO
-import qualified Dhall.ETags      as ETags
+import qualified Dhall.Tags       as Tags
 import qualified Dhall.Test.Util  as Test.Util
 import qualified Test.Tasty       as Tasty
 import qualified Test.Tasty.HUnit as Tasty.HUnit
 import qualified Turtle
 
-etagsDirectory :: FilePath
-etagsDirectory = "./tests/etags"
+tagsDirectory :: FilePath
+tagsDirectory = "./tests/tags"
 
 getTests :: IO TestTree
 getTests = do
-    etagsTests <- Test.Util.discover (Turtle.chars <* ".dhall") etagsTest (Turtle.lstree etagsDirectory)
+    tagsTests <- Test.Util.discover (Turtle.chars <* ".dhall") tagsTest (Turtle.lstree tagsDirectory)
 
-    let testTree = Tasty.testGroup "etags tests" [ etagsTests, etagsDirTest ]
+    let testTree = Tasty.testGroup "tags tests" [ tagsTests, tagsDirTest ]
 
     return testTree
 
-etagsTest :: Text -> TestTree
-etagsTest prefix =
+tagsTest :: Text -> TestTree
+tagsTest prefix =
     Tasty.HUnit.testCase (Text.unpack prefix) $ do
         let inputFile  = Text.unpack (prefix <> ".dhall")
         let outputFile = Text.unpack (prefix <> ".tags")
 
-        actualTags <- fixPathSeparators <$> ETags.generate (InputFile inputFile) Nothing False
+        actualTags <- fixPathSeparators <$> Tags.generate (InputFile inputFile) Nothing False
 
         expectedTags <- Text.IO.readFile outputFile
 
@@ -42,14 +42,14 @@ etagsTest prefix =
 
         Tasty.HUnit.assertEqual message expectedTags actualTags
 
-etagsDirTest :: TestTree
-etagsDirTest =
+tagsDirTest :: TestTree
+tagsDirTest =
     Tasty.HUnit.testCase "all" $ do
-        let outputFile = Text.unpack . Turtle.format Turtle.fp $ etagsDirectory Turtle.</> "all.tags"
+        let outputFile = Text.unpack . Turtle.format Turtle.fp $ tagsDirectory Turtle.</> "all.tags"
 
         actualTags <- fmap fixPathSeparators
-                      (ETags.generate
-                          (InputFile (Text.unpack . Turtle.format Turtle.fp $ etagsDirectory))
+                      (Tags.generate
+                          (InputFile (Text.unpack . Turtle.format Turtle.fp $ tagsDirectory))
                           (Just [".dhall"])
                           False)